Converted button menu methods to use floating point values instead of a Point object.

+++ EditorLib/src/shapes/PolygonBuilder.java
@@ -8,7 +8,7 @@ final public class PolygonBuilder implements Library {
 		
 	}
 
-	public static Polygon makeBox(int width, int height) {
+	public static Polygon makeBox(float width, float height) {
 		Polygon polygon = new Polygon(4);
 		polygon.addPoint(0, 0);
 		polygon.addPoint(width, 0);
@@ -17,7 +17,7 @@ final public class PolygonBuilder implements Library {
 		return polygon;
 	}
 	
-	public static Polygon makeUpArrow(int width, int height) {
+	public static Polygon makeUpArrow(float width, float height) {
 		Polygon polygon = new Polygon(3);
 		polygon.addPoint(0, 0);
 		polygon.addPoint(-width/2, height);
@@ -25,7 +25,7 @@ final public class PolygonBuilder implements Library {
 		return polygon;
 	}
 	
-	public static Polygon makeDownArrow(int width, int height) {
+	public static Polygon makeDownArrow(float width, float height) {
 		Polygon polygon = new Polygon(3);
 		polygon.addPoint(0, 0);
 		polygon.addPoint(-width/2, -height);
@@ -33,7 +33,7 @@ final public class PolygonBuilder implements Library {
 		return polygon;
 	}
 	
-	public static Polygon makeLeftArrow(int width, int height) {
+	public static Polygon makeLeftArrow(float width, float height) {
 		Polygon polygon = new Polygon(3);
 		polygon.addPoint(0, 0);
 		polygon.addPoint(width, -height/2);
@@ -41,7 +41,7 @@ final public class PolygonBuilder implements Library {
 		return polygon;
 	}
 	
-	public static Polygon makeRightArrow(int width, int height) {
+	public static Polygon makeRightArrow(float width, float height) {
 		Polygon polygon = new Polygon(3);
 		polygon.addPoint(0, 0);
 		polygon.addPoint(-width, -height/2);
